move listener start from listening.rs to lib.rs
